Move Try{Update,Patch} to extensions library
