+++ extensions/pkg/controller/patch_test.go
@@ -12,38 +12,40 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package kubernetes
+package controller
 
 import (
 	"context"
 	"encoding/json"
 
+	apierrors "k8s.io/apimachinery/pkg/api/errors"
+	"k8s.io/apimachinery/pkg/runtime/schema"
+	"sigs.k8s.io/controller-runtime/pkg/client"
+
 	"github.com/gardener/gardener/pkg/apis/core/v1beta1"
 	extensionsv1alpha1 "github.com/gardener/gardener/pkg/apis/extensions/v1alpha1"
+	"github.com/gardener/gardener/pkg/utils/kubernetes"
 
 	. "github.com/onsi/ginkgo"
-
 	. "github.com/onsi/gomega"
-	apierrors "k8s.io/apimachinery/pkg/api/errors"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/apimachinery/pkg/runtime"
-	"k8s.io/apimachinery/pkg/runtime/schema"
 	"k8s.io/client-go/util/retry"
-	"sigs.k8s.io/controller-runtime/pkg/client"
 	"sigs.k8s.io/controller-runtime/pkg/client/fake"
 )
 
-var _ = Describe("#tryUpdate", func() {
+var _ = Describe("#tryPatch", func() {
 	It("should set state to obj, when conflict occurs", func() {
 		s := runtime.NewScheme()
 		Expect(extensionsv1alpha1.AddToScheme(s)).NotTo(HaveOccurred())
 		objInFakeClient := newInfraObj()
+		objInFakeClient.SetResourceVersion("1")
 		objInFakeClient.Status.Conditions = []v1beta1.Condition{
 			{Type: "Health", Reason: "reason", Message: "messages", Status: "status", LastUpdateTime: metav1.Now()},
 		}
 
 		c := fake.NewClientBuilder().WithScheme(s).WithObjects(objInFakeClient).Build()
-		infraObj := newInfraObj()
+		infraObj := objInFakeClient.DeepCopy()
 		transform := func() error {
 			infraState, _ := json.Marshal(state{"someState"})
 			infraObj.GetExtensionStatus().SetState(&runtime.RawExtension{Raw: infraState})
@@ -55,38 +57,19 @@ var _ = Describe("#tryUpdate", func() {
 			client:                c,
 		}
 
-		tryUpdateErr := tryUpdate(context.TODO(), retry.DefaultRetry, c, infraObj, u.updateFunc, transform)
-		Expect(tryUpdateErr).NotTo(HaveOccurred())
+		tryPatchErr := tryPatch(context.Background(), retry.DefaultRetry, c, infraObj, u.patchFunc, transform)
+		Expect(tryPatchErr).NotTo(HaveOccurred())
 
 		objFromFakeClient := &extensionsv1alpha1.Infrastructure{}
-		err := c.Get(context.TODO(), Key("infraNamespace", "infraName"), objFromFakeClient)
+		err := c.Get(context.Background(), kubernetes.Key("infraNamespace", "infraName"), objFromFakeClient)
 		Expect(err).NotTo(HaveOccurred())
 		Expect(objFromFakeClient).To(Equal(infraObj))
 	})
 })
 
-type state struct {
-	Name string `json:"name"`
-}
-
-func newInfraObj() *extensionsv1alpha1.Infrastructure {
-	return &extensionsv1alpha1.Infrastructure{
-		ObjectMeta: metav1.ObjectMeta{
-			Name:      "infraName",
-			Namespace: "infraNamespace",
-		},
-	}
-}
-
-type conflictErrManager struct {
-	conflictsBeforeUpdate int
-	conflictsOccured      int
-	client                client.Client
-}
-
-func (c *conflictErrManager) updateFunc(ctx context.Context, obj client.Object, o ...client.UpdateOption) error {
+func (c *conflictErrManager) patchFunc(ctx context.Context, obj client.Object, patch client.Patch, o ...client.PatchOption) error {
 	if c.conflictsBeforeUpdate == c.conflictsOccured {
-		return c.client.Status().Update(ctx, obj, o...)
+		return c.client.Status().Patch(ctx, obj, patch, o...)
 	}
 
 	c.conflictsOccured++
